@@ -21,27 +21,30 @@
   },
   "typings": "dist/index.d.ts",
   "dependencies": {
-    "fuse.js": "^6.4.6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "styled-components": "^5.3.1"
+    "fuse.js": "^6.6.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "styled-components": "^5.3.6"
   },
   "devDependencies": {
-    "@commitlint/cli": "^13.2.1",
-    "@commitlint/config-conventional": "^13.2.0",
-    "@release-it/conventional-changelog": "^3.0.0",
-    "@types/node": "^16.11.1",
-    "@types/react": "^17.0.30",
-    "@types/react-dom": "^17.0.9",
-    "@types/styled-components": "^5.1.15",
-    "@vitejs/plugin-react": "^1.0.4",
-    "husky": "^7.0.0",
-    "lint-staged": "^11.0.0",
-    "prettier": "^2.2.1",
-    "release-it": "^14.2.2",
-    "tsup": "^5.4.1",
-    "typescript": "^4.4.4",
-    "vite": "^2.6.7"
+    "@commitlint/cli": "^17.4.2",
+    "@commitlint/config-conventional": "^17.4.2",
+    "@release-it/conventional-changelog": "^5.1.1",
+    "@types/node": "^18.13.0",
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "@types/styled-components": "^5.1.26",
+    "@vitejs/plugin-react": "^3.1.0",
+    "husky": "^8.0.3",
+    "lint-staged": "^13.1.1",
+    "prettier": "^2.8.3",
+    "release-it": "^15.6.0",
+    "tsup": "^6.6.0",
+    "typescript": "^4.9.5",
+    "vite": "^4.1.1"
+  },
+  "resolutions": {
+    "@types/react": "^18.0.21"
   },
   "exports": {
     ".": {
