@@ -1,8 +1,11 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import ReactSearchBox from "../src";
 
-ReactDOM.render(
+const container = document.getElementById("root");
+const root = createRoot(container!);
+
+root.render(
   <React.StrictMode>
     <ReactSearchBox
       placeholder="Search for John, Jane or Mary"
@@ -37,6 +40,5 @@ ReactDOM.render(
       leftIcon={<>🎨</>}
       iconBoxSize="48px"
     />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
